backend/osint_modules/email_lookup.py
"""
Email Lookup Module
"""

import re
import requests
import dns.resolver
from typing import Dict, List
import socket
import smtplib

class EmailLookup:
    """Email address lookup and validation"""
    
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
    
    def lookup(self, email: str) -> Dict:
        """Lookup email information"""
        results = {
            'email': email,
            'valid_format': False,
            'domain': None,
            'domain_info': {},
            'breach_data': [],
            'social_profiles': []
        }
        
        # Validate email format
        email_pattern = r'^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
        if re.match(email_pattern, email):
            results['valid_format'] = True
            results['domain'] = email.split('@')[1]
            results['domain_info'] = self.get_domain_info(results['domain'])
            results['breach_data'] = self.check_breaches(email)
            results['social_profiles'] = self.search_social_profiles(email)
        
        return results
    
    def search_by_name(self, name: str) -> List[Dict]:
        """Search for email addresses by name"""
        # This would typically use email finder APIs
        results = []
        try:
            # Common email patterns
            name_parts = name.lower().split()
            if len(name_parts) >= 2:
                patterns = [
                    f"{name_parts[0]}.{name_parts[-1]}",
                    f"{name_parts[0]}{name_parts[-1]}",
                    f"{name_parts[0][0]}{name_parts[-1]}",
                ]
                results.append({
                    'suggested_patterns': patterns,
                    'note': 'Use email finder APIs for actual results'
                })
        except Exception as e:
            print(f"Email search error: {e}")
        return results
    
    def get_domain_info(self, domain: str) -> Dict:
        """Get domain information"""
        info = {
            'domain': domain,
            'mx_records': [],
            'has_mx': False
        }
        
        try:
            mx_records = dns.resolver.resolve(domain, 'MX')
            info['has_mx'] = True
            for mx in mx_records:
                info['mx_records'].append(str(mx.exchange))
        except Exception as e:
            info['error'] = str(e)
        
        return info
    
    def check_breaches(self, email: str) -> List[Dict]:
        """Check if email was in data breaches"""
        # This would use HaveIBeenPwned API or similar
        return []
    
    def search_social_profiles(self, email: str) -> List[Dict]:
        """Search for social media profiles by email"""
        # This would use social media APIs
        return []

