backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
Pillow==10.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1
googlemaps==4.10.0
phonenumbers==8.13.22
email-validator==2.1.0
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
pytesseract==0.3.10
whois==0.9.16
dnspython==2.4.2
shodan==1.31.0
python-whois==0.8.0
social-analyzer==0.45.0
tweepy==4.14.0
instaloader==4.10.3

